requirements.txt
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
scipy>=1.10.0
numpy>=1.23.0
Pillow>=9.4.0
trimesh>=4.0.0
accelerate>=0.20.0
